+++ library/components/molecules/canvas-reveal.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { cn } from "../lib/utils";
+import { cn } from "@/utils";
 import { Canvas, useFrame, useThree } from "@react-three/fiber";
 import React, { useMemo, useRef } from "react";
 import * as THREE from "three";
