+++ library/components/molecules/text-generate.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { useEffect } from "react";
 import { motion, stagger, useAnimate } from "framer-motion";
-import { cn } from "../lib/utils";
+import { cn } from "@/utils";
 
 export const TextGenerateEffect = ({
   words,
